update sound for action, update Algorithm::CalculateExperience, fix few bug
git-svn-id: https://fheroes2.svn.sourceforge.net/svnroot/fheroes2/trunk@489 4f904287-c015-0410-838d-95d3a2656228

+++ fheroes2/src/fheroes2/algorithm/algorithm.h
@@ -28,9 +28,9 @@ class Castle;
 
 namespace Algorithm
 {
-    u16 CalculateExperience(const Army::army_t & army);
-    u16 CalculateExperience(const Heroes & hero);
-    u16 CalculateExperience(const Castle &castle);
+    u32 CalculateExperience(const Army::army_t & army);
+    u32 CalculateExperience(const Heroes & hero);
+    u32 CalculateExperience(const Castle &castle);
 };
 
 #endif
